style:rename folders
