+++ marvel/src/routes/SingleSeriePage/SingleSeriePage.module.scss
